Fixed changed types and upgrade the login page

+++ ViewModels/ReporteRequest.cs
@@ -51,7 +51,7 @@ public class ReporteRequest
 
     [Display(Name = "Estatus")]
     [Range(1, 10)]
-    public decimal? IdEstatus { get; set; }
+    public int? IdEstatus { get; set; }
 
     [StringLength(4000)]
     public string? Referencias { get; set; }
